Enable warnings as errors for first-party code; fix resulting findings
Build system (Premake/ProjectUtil.lua): projects outside the dependency
group now build with FatalCompileWarnings; dependency projects are
exempt and dependency headers are consumed as external includes.

Fixes surfaced by the sweep:
- ActionStack.h: BaseAction is deleted polymorphically but had no
  virtual destructor - UB; add one
- AnimationUtil.cpp: boneIndex (u16) was compared against the i16 -1
  InvalidBoneID sentinel, which can never match
  (-Wtautological-constant-out-of-range-compare); compare against the
  u16 representation
- UnitUtil.cpp, UnitHandler.cpp: add default cases to non-exhaustive
  switches over animation and unit class enums (-Wswitch), preserving
  the existing fallthrough behavior

+++ Source/Game-Lib/Game-Lib/Util/AnimationUtil.cpp
@@ -399,7 +399,7 @@ namespace Util::Animation
         u32 numBoneMatrices = static_cast<u32>(animationData.boneTransforms.size());
         u16 boneIndex = GetBoneIndexFromKeyBoneID(modelInfo, bone);
 
-        if (boneIndex == ::Animation::Defines::InvalidBoneID || boneIndex >= numBoneMatrices)
+        if (boneIndex == static_cast<u16>(::Animation::Defines::InvalidBoneID) || boneIndex >= numBoneMatrices)
             return nullptr;
 
         return GetBoneMatrixRaw(animationData, boneIndex);
